requirements.txt
livekit
livekit-agents[openai,deepgram,silero]
asyncpg
aiohttp
python-dotenv
psycopg2-binary